+++ src/browser-util.js
@@ -21,10 +21,6 @@
    var root = this;
    var exports = root.jsPlumbUtil;
 
-   exports.ieVersion = /MSIE\s([\d.]+)/.test(navigator.userAgent) ? (new Number(RegExp.$1)) : -1;
-
-   exports.oldIE = exports.ieVersion > -1 && exports.ieVersion < 9;
-
    exports.matchesSelector = function(el, selector, ctx) {
        ctx = ctx || el.parentNode;
        var possibles = ctx.querySelectorAll(selector);